Small program to calculate average tips.
This program takes input from a user to keep track of how many hours
they work in a shift and how much they make in tips, pushes those
values into their respective arrays then calculates the average for
tips and how much the user makes in an hour.  it is currently
functional, but I’d like feedback on how to make the code shorter and
more efficient.